+++ problems/tool_creator_challenge/sandbox/environment.py
@@ -72,5 +72,5 @@ class ToolCreatorEnvironment:
     # Scoring
     # ------------------------------------------------------------------
 
-    def compute_score(self) -> int:
+    def compute_score(self) -> float:
         return 100 if self._success else 0
